+++ ccramic/utils/db.py
@@ -1,4 +1,5 @@
 import pandas as pd
+from ccramic.io.session import JSONSessionDocument
 
 def match_db_config_to_request_str(db_config_list: list, db_config_selection: str):
     """
@@ -62,20 +63,13 @@ def preview_dataframe_from_db_config_list(config_list):
 
 def format_blend_config_document_for_insert(user, config_name, blend_dict, selected_channel_list, global_apply_filter,
                                             global_filter_type, global_filter_val, global_filter_sigma,
+                                            data_selection: str=None, cluster_assignments: dict=None,
                                             aliases: dict=None):
     """
     Format a mongoDB document from a session config that can be posted to the `blend_config` collection
     in the `ccramic-db` mongoDB database
     """
-    # TODO: add the alias to the config
-    if aliases is not None:
-        for key in blend_dict.keys():
-            if key in aliases.keys():
-                blend_dict[key]['alias'] = aliases[key]
-    return {"user": user,
-        "name": config_name,
-        "channels": blend_dict,
-        "config": {"blend": selected_channel_list, "filter": {"global_apply_filter": global_apply_filter,
-                                                                  "global_filter_type": global_filter_type,
-                                                                  "global_filter_val": global_filter_val,
-                                                              "global_filter_sigma": global_filter_sigma}}}
+    return JSONSessionDocument("db", user, config_name, blend_dict, selected_channel_list, global_apply_filter,
+                                         global_filter_type, global_filter_val, global_filter_sigma,
+                                         data_selection, cluster_assignments, aliases).get_document()
+